% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castlow_str.R
\name{castlow.str}
\alias{castlow.str}
\title{Cast of a character string}
\usage{
castlow.str(st)
}
\arguments{
\item{st}{character string}
}
\value{
Returns the ASCII//TRANSLIT transcription of the word \code{st},
without accents, spaces and in lowercase letters.
}
\description{
The \code{castlow.str} function converts a word to lowercase,
without accents and spaces.
}
\examples{
castlow.str (st = c("Right eye", "Left_Lung", "Right-Lung"))
}
\seealso{
\link[espadon]{castup.str}.
}
