% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castup_str.R
\name{castup.str}
\alias{castup.str}
\title{Cast of a character string}
\usage{
castup.str(st)
}
\arguments{
\item{st}{character string}
}
\value{
Returns the ASCII//TRANSLIT transcription of the word \code{st},
without accents, spaces and in capitals.
}
\description{
The \code{castup.str} function converts a word to upper case,
without accents and spaces.
}
\examples{
castup.str (st = c("Right eye", "Left_Lung", "Right-Lung"))
}
\seealso{
\link[espadon]{castlow.str}.
}
