% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_update_roiinfo.R
\name{struct.update_roiinfo}
\alias{struct.update_roiinfo}
\title{Update roi.info}
\usage{
struct.update_roiinfo(struct)
}
\arguments{
\item{struct}{espadon object of class struct.}
}
\value{
Returns espadon object of class struct with updated roi.info element.
}
\description{
The \code{struct.update_roiinfo} function updates the roi.info
element in the swordfish object of the struct class when point coordinates
have been modified in the roi.data element.
}
