% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_dicom_raw.R
\name{toy.dicom.raw}
\alias{toy.dicom.raw}
\title{toy DICOM raw data}
\usage{
toy.dicom.raw()
}
\value{
Returns the raw data of a dummy DICOM file of rtplan modality.
}
\description{
The \code{toy.dicom.raw} loads raw data from a dummy DICOM file.
It is used for the test.
}
\examples{
toy.dicom.raw ()
}
