% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_product.R
\name{vector.product}
\alias{vector.product}
\title{Vector product of two vectors}
\usage{
vector.product(v1, v2)
}
\arguments{
\item{v1}{Vector of x, y, z coordinates}

\item{v2}{Vector of x, y, z coordinates}
}
\value{
Returns the x, y, z coordinates of the vector product of v1 and v2
}
\description{
Vector product of two vectors
}
\examples{
vector.product(c (1, 0, 0), c (0, 1, 0))

}
